+++ packages/stream_chat_flutter/lib/src/attachment/builder/attachment_widget_builder.dart
@@ -67,8 +67,11 @@ abstract class StreamAttachmentWidgetBuilder {
     ShapeBorder? shape,
     EdgeInsetsGeometry padding = const EdgeInsets.all(4),
     StreamAttachmentWidgetTapCallback? onAttachmentTap,
+    List<StreamAttachmentWidgetBuilder>? customAttachmentBuilders,
   }) {
     return [
+      ...?customAttachmentBuilders,
+
       // Handles a mix of image, gif, video, url and file attachments.
       MixedAttachmentBuilder(
         padding: padding,